+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 ## v0.3.0
 - Added Celestium, Unholy, and Space Dust blocks
 - Can now convert Sand of Time directly into Celestium, Unholy, and Space Dust blocks
+- Added REI recipes for conversion between Sand of Time and the 3 dusts
 - Added Unholy Grass Block (randomly spawns mobs when below 8 light level)
+- Added button toggles for Wormhole Transporter to enable/disable Items, Fluids, and Energy
 - Added a button in the Divine Replicator to toggle between Exact and Similar spawns
 - Fixed bug that was preventing Unholy Silencer from receiving unholy dust
+- Fixed bug with rendering the correct fluid amount in the Wormhole Transporter
 
 ---
 ## v0.2.1